Move metrics to a separate unit, avoid unnecessary enum
